Revert "refactor: address all 17 code smells in app/"
