+++ .env.example
@@ -44,13 +44,6 @@ APP_ENV=development          # development | staging | production
 LOG_LEVEL=INFO
 RETRIEVER_K=5
 
-# ── CORS ──────────────────────────────────────
-# Comma-separated exact origins allowed by the FastAPI CORS middleware.
-# Defaults cover local Vite (5173) and CRA (3000) dev servers.
-CORS_ALLOWED_ORIGINS=http://localhost:5173,http://localhost:3000
-# Regex applied in addition to CORS_ALLOWED_ORIGINS. Set to empty to disable.
-CORS_ALLOWED_ORIGIN_REGEX=https://.*\.vercel\.app$
-
 # ── Ingestion ─────────────────────────────────
 # Local path or remote URL used by the ingest pipeline
 DATA_SOURCE_PATH=./data/cities.jsonl
