--- app/routes.py
@@ -1,109 +1,0 @@
-from __future__ import annotations
-
-import logging
-import time
-import uuid
-from typing import Any
-
-from fastapi import APIRouter, Depends, HTTPException, Request
-from psycopg2.extensions import connection
-
-from .db import get_db
-from .errors import status_for_invoke_error
-from .schemas import (
-    RecommendationRequest,
-    RecommendationResponse,
-    RecommendationSource,
-)
-
-logger = logging.getLogger(__name__)
-router = APIRouter()
-db_connection_dependency = Depends(get_db)
-
-
-def serialize_sources(source_documents: list[Any]) -> list[RecommendationSource]:
-    sources: list[RecommendationSource] = []
-    for document in source_documents:
-        metadata = getattr(document, "metadata", {}) or {}
-        sources.append(
-            RecommendationSource(
-                name=metadata.get("name"),
-                rating=metadata.get("rating"),
-                address=metadata.get("address"),
-                similarity=metadata.get("similarity"),
-            )
-        )
-    return sources
-
-
-@router.get("/root")
-def root() -> dict[str, str]:
-    return {"message": "Welcome!"}
-
-
-@router.get("/health")
-def health(request: Request) -> dict[str, str]:
-    model_config = getattr(request.app.state, "active_model_config", None)
-    if model_config is None:
-        startup_error = getattr(request.app.state, "startup_error", None) or "unknown"
-        raise HTTPException(
-            status_code=503,
-            detail={"status": "degraded", "error": startup_error},
-        )
-
-    return {
-        "status": "ok",
-        "llm_provider": model_config.llm_provider,
-        "chat_model": model_config.chat_model,
-    }
-
-
-@router.get("/health/db")
-def health_db(conn: connection = db_connection_dependency) -> dict[str, str]:
-    with conn.cursor() as cur:
-        cur.execute("SELECT 1")
-        _ = cur.fetchone()
-    return {"status": "ok"}
-
-
-@router.post("/predict", response_model=RecommendationResponse)
-def predict(request_body: RecommendationRequest, request: Request) -> RecommendationResponse:
-    request_id = uuid.uuid4().hex[:8]
-    start = time.perf_counter()
-    logger.info("predict[%s] start: query_len=%d", request_id, len(request_body.query))
-
-    rag_chain = getattr(request.app.state, "rag_chain", None)
-    if rag_chain is None:
-        logger.warning("predict[%s] rag_chain not loaded", request_id)
-        raise HTTPException(status_code=500, detail="RAG chain is not loaded.")
-
-    try:
-        result = rag_chain.invoke({"query": request_body.query})
-    except Exception as exc:
-        status_code, message_prefix = status_for_invoke_error(exc)
-        logger.exception(
-            "predict[%s] failed: status=%d type=%s",
-            request_id,
-            status_code,
-            type(exc).__name__,
-        )
-        raise HTTPException(
-            status_code=status_code,
-            detail=f"{message_prefix} (request_id={request_id}).",
-        ) from exc
-
-    response_text = result.get("result") or result.get("response") or ""
-    source_documents = result.get("source_documents") or []
-
-    elapsed_ms = (time.perf_counter() - start) * 1000
-    logger.info(
-        "predict[%s] ok: sources=%d elapsed_ms=%.0f",
-        request_id,
-        len(source_documents),
-        elapsed_ms,
-    )
-
-    return RecommendationResponse(
-        response=response_text,
-        sources=serialize_sources(source_documents),
-    )
